+++ Code-Zero/Global/Service/UserSettingService.swift
@@ -45,4 +45,34 @@ class UserSettingService {
             }
         }
     }
+
+    public func requestUserInfo(token: String,
+                                completion: @escaping (NetworkResult<SettingData>) -> Void
+    ) {
+        service.request(APITarget.userInfo(token: token)) { result in
+            switch result {
+            case .success(let response):
+                do {
+                    let decoder = JSONDecoder()
+                    let body = try decoder.decode(
+                        GenericResponse<SettingData>.self,
+                        from: response.data
+                    )
+                    guard let data = body.data else { return }
+                    completion(.success(data))
+                } catch let error {
+                    debugPrint(error)
+                }
+            case .failure(let error):
+                // 에러 처리 부분
+                guard let response = error.response else { return }
+                if response.statusCode == 500 {
+                    completion(.networkFail)
+                    return
+                }
+                completion(.serverErr)
+                debugPrint(error)
+            }
+        }
+    }
 }
